# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Tesis Guvenlik & Operasyon API
  version: "0.1.0"
  description: |
    Multi-tenant tesis guvenlik & operasyon SaaS — v0 sozlesmesi.
    Backend (FastAPI) ile Mobil (Flutter) ve Panel (Next.js) gelistiricilerinin
    paralel calismasi icin tek dogruluk kaynagidir.

    Konvansiyonlar (detay: /contracts/README.md):
      * Tum zamanlar ISO8601 UTC (`...Z`). tenant.timezone ile yorumlanir.
      * Hata zarfi: `{ "error": { "code": "...", "message": "..." } }`.
      * Sayfalama: `limit` (varsayilan 50, max 200) + `offset`. Liste yanitlari
        `PageMeta` icerir.
      * Auth: JWT Bearer (bkz. /contracts/auth.md). Tenant her zaman token'dan
        cikarilir — istemci tenant_id gondermez/secemez.
      * `POST /scans` icin `Idempotency-Key` header zorunlu.
servers:
  # NOT: base path /v0 KALDIRILDI — gercek backend endpoint'leri /v0 ONEKI OLMADAN
  # sunuyor (orn. /auth/login, /scans). Base = host:port.
  - url: https://api.example.com
    description: production
  - url: http://localhost:8000
    description: yerel gelistirme
...
security:
  - bearerAuth: []
paths:
...
  /units/{id}/dues:
    parameters:
      - $ref: "#/components/parameters/PathId"
    get:
      tags: [aidat]
      summary: Dairenin borc durumu (tahakkuk - odeme = bakiye)
      responses:
        "200":
          description: Borc durumu
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnitDuesStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /dues/assessments:
    get:
      tags: [aidat]
      summary: Tahakkuk listesi
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: unit_id
          in: query
          schema: { type: string, format: uuid }
        - name: donem
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Liste
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/DuesAssessment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [aidat]
      summary: Tahakkuk olustur (tek daire veya toplu donem) (admin)
      description: |
        `unit_id` verilirse TEK daire (ayni daire+donem ikinci kez → 409). `unit_id`
        yoksa TOPLU: `unit_ids` (verilmezse tum aktif daireler); mevcut donemler atlanir.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DuesAssessmentCreate" }
      responses:
        "201":
          description: "Olusturuldu (toplu modda created + atlanan)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DuesAssessmentResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Tek daire modunda ayni donem zaten var
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /dues/payments:
    get:
      tags: [aidat]
      summary: Odeme listesi
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: unit_id
          in: query
          schema: { type: string, format: uuid }
        - name: donem
          in: query
          description: "'YYYY-MM' — donem bazli rapor filtresi (payment.donem uzerinden)"
          schema: { type: string }
      responses:
        "200":
          description: Liste
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    required: [items]
                    properties:
                      items:
                        type: array
                        items: { $ref: "#/components/schemas/DuesPayment" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [aidat]
      summary: Manuel odeme kaydi (admin)
      description: |
        Gercek tahsilat YOK (soyut). Idempotency-Key ZORUNLU (cift kayit korumasi).
        kaydeden token'dan. Bakiye = toplam tahakkuk - basarili odemeler.

        DONEM ATFI (panel raporu): `donem` ('YYYY-MM') acikca verilebilir;
        verilmezse `assessment_id` varsa ONUN doneminden turer; ikisi de yoksa
        NULL kalir (serbest odeme). Rapor `GET /dues/payments?donem=` ile
        serbest odemeleri de doneme atfedebilir.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DuesPaymentCreate" }
      responses:
        "201":
          description: Odeme kaydedildi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DuesPayment" }
        "200":
          description: Idempotent tekrar — mevcut kayit
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DuesPayment" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Ayni Idempotency-Key farkli govde ile gonderildi
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /me/dues:
    get:
      tags: [aidat]
      summary: Sakinin kendi dairelerinin borc durumu (resident)
      responses:
        "200":
          description: Borc durumu listesi
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/UnitDuesStatus" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ================================================================= #
  # ODEME WEBHOOK (public ama imza-korumali)
  # ================================================================= #
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. `Authorization: Bearer <token>`. Bkz. /contracts/auth.md"
  parameters:
    Limit:
      name: limit
      in: query
      description: Sayfa boyutu (varsayilan 50, max 200)
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      description: Atlanan kayit sayisi
      schema: { type: integer, minimum: 0, default: 0 }
    PathId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Istemci tarafindan uretilen benzersiz anahtar (UUID onerilir)
      schema: { type: string, minLength: 8, maxLength: 200 }
  responses:
    BadRequest:
      description: Gecersiz istek
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Kimlik dogrulama gerekli / gecersiz token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Rol bu islem icin yetkili degil
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Kayit bulunamadi (tenant kapsaminda)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    ValidationError:
      description: Govde dogrulama hatasi
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Makine-okunabilir hata kodu
              example: validation_error
            message:
              type: string
              description: Insan-okunabilir aciklama
              example: "baslangic_saat bitis_saat'ten once olmali"
            details:
              type: array
              description: Alan bazli dogrulama hatalari (opsiyonel)
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    Page:
      type: object
      required: [meta]
      properties:
        meta: { $ref: "#/components/schemas/PageMeta" }
    PageMeta:
      type: object
      required: [limit, offset, total]
      properties:
        limit: { type: integer, example: 50 }
        offset: { type: integer, example: 0 }
        total:
          type: integer
          description: Filtreye uyan toplam kayit sayisi
          example: 137
...
    DuesYontem:
      type: string
      enum: [elden, havale, kart, diger]
    DuesDurum:
      type: string
      enum: [basarili, bekliyor, iptal]
...
    DuesAssessment:
      type: object
      required: [id, unit_id, donem, tutar_kurus, created_at]
      properties:
        id: { type: string, format: uuid }
        unit_id: { type: string, format: uuid }
        donem: { type: string, example: "2026-06" }
        tutar_kurus:
          type: integer
          minimum: 1
          description: "KURUS (minor units); para asla float degil"
        son_odeme_tarihi: { type: string, format: date, nullable: true }
        aciklama: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    DuesAssessmentCreate:
      type: object
      required: [donem, tutar_kurus]
      properties:
        unit_id:
          type: string
          format: uuid
          nullable: true
          description: "Verilirse TEK daire; yoksa toplu"
        unit_ids:
          type: array
          nullable: true
          description: "Toplu modda hedef daireler; yoksa tum aktif daireler"
          items: { type: string, format: uuid }
        donem: { type: string, example: "2026-06" }
        tutar_kurus: { type: integer, minimum: 1 }
        son_odeme_tarihi: { type: string, format: date, nullable: true }
        aciklama: { type: string, nullable: true }
    DuesAssessmentResult:
      type: object
      required: [created, atlanan]
      properties:
        created:
          type: array
          items: { $ref: "#/components/schemas/DuesAssessment" }
        atlanan:
          type: integer
          description: "Zaten var olan (donem) tahakkuk sayisi (toplu modda atlanan)"
    DuesPayment:
      type: object
      required: [id, unit_id, tutar_kurus, odeme_zamani, yontem, durum, kaydeden_user_id, idempotency_key, created_at]
      properties:
        id: { type: string, format: uuid }
        unit_id: { type: string, format: uuid }
        assessment_id: { type: string, format: uuid, nullable: true }
        tutar_kurus: { type: integer, minimum: 1 }
        odeme_zamani: { type: string, format: date-time }
        donem:
          type: string
          nullable: true
          example: "2026-06"
          description: "Odemenin atfedildigi donem; serbest odemede NULL olabilir"
        yontem: { $ref: "#/components/schemas/DuesYontem" }
        durum: { $ref: "#/components/schemas/DuesDurum" }
        makbuz_no: { type: string, nullable: true }
        provider: { type: string, nullable: true, description: "manual|iyzico|paytr" }
        provider_ref: { type: string, nullable: true, description: "Saglayici referansi (webhook bunu kullanir)" }
        odeme_url:
          type: string
          nullable: true
          description: "Yalniz kart akisi init yanitinda: saglayici odeme sayfasi URL'i (kalici degil)"
        kaydeden_user_id: { type: string, format: uuid }
        idempotency_key: { type: string }
        created_at: { type: string, format: date-time }
    DuesPaymentCreate:
      type: object
      required: [unit_id, tutar_kurus, yontem]
      properties:
        unit_id: { type: string, format: uuid }
        assessment_id: { type: string, format: uuid, nullable: true }
        tutar_kurus: { type: integer, minimum: 1 }
        yontem: { $ref: "#/components/schemas/DuesYontem" }
        makbuz_no: { type: string, nullable: true }
        odeme_zamani:
          type: string
          format: date-time
          nullable: true
          description: "Verilmezse sunucu now()"
        donem:
          type: string
          nullable: true
          example: "2026-06"
          description: "Verilmezse assessment'tan turer; o da yoksa NULL kalir"
    UnitDuesStatus:
      type: object
      required: [unit_id, no, toplam_tahakkuk_kurus, toplam_odenen_kurus, bakiye_kurus]
      properties:
        unit_id: { type: string, format: uuid }
        no: { type: string }
        toplam_tahakkuk_kurus: { type: integer }
        toplam_odenen_kurus:
          type: integer
          description: "Yalniz durum='basarili' odemeler"
        bakiye_kurus:
          type: integer
          description: "toplam_tahakkuk - toplam_odenen (pozitif = borc)"
        assessments:
          type: array
          items: { $ref: "#/components/schemas/DuesAssessment" }
        payments:
          type: array
          items: { $ref: "#/components/schemas/DuesPayment" }